Commit du CRUD des nouveaux package

+++ src/main/java/fr/nathan/seances/SeanceRepository.java
@@ -2,5 +2,5 @@ package fr.nathan.seances;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeanceRepository extends JpaRepository<Seance, Integer> {
+public interface SeanceRepository extends JpaRepository<Seance, Long> {
 }
